# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KBJ API",
    "description": "KBJ 로그인 등급 API(사용자 1명). 모든 데이터 응답은 봉투(source·as_of·quality·notes·generated_at·data), 금액은 원 단위 정수. 외부 원천을 부르지 않는다(DB·Redis 읽기만).",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/etf/flows": {
      "get": {
        "tags": [
          "etf"
        ],
        "summary": "Flows",
        "operationId": "flows_api_etf_flows_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "in",
                "out",
                "value",
                "indiv",
                "foreign",
                "inst"
              ],
              "type": "string",
              "default": "in",
              "title": "Mode"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EtfType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                1,
                5,
                20
              ],
              "type": "integer",
              "default": 5,
              "title": "Period"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_EtfFlows_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/etf/types": {
      "get": {
        "tags": [
          "etf"
        ],
        "summary": "Types",
        "operationId": "types_api_etf_types_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                1,
                5,
                20
              ],
              "type": "integer",
              "default": 5,
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_EtfTypes_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/etf/premium": {
      "get": {
        "tags": [
          "etf"
        ],
        "summary": "Premium",
        "operationId": "premium_api_etf_premium_get",
        "parameters": [
          {
            "name": "basis",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "nav",
                "inav"
              ],
              "type": "string",
              "default": "nav",
              "title": "Basis"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_EtfPremium_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/etf/holdings/changes": {
      "get": {
        "tags": [
          "etf"
        ],
        "summary": "Holding Changes",
        "operationId": "holding_changes_api_etf_holdings_changes_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(NEW|DROP|IN10|OUT10|ADD|CUT)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "issuer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issuer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_EtfHoldingChanges_"
                }
              }
            }
          },
          "401": {
            "description": "세션 없음·만료",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "아직 없음(작업·예정 시각)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoDataBody"
                }
              }
            }
          },
          "422": {
            "description": "쿼리 검증 실패(입력값은 싣지 않는다)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "저장소·Redis 장애",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Check3": {
        "properties": {
          "net_asset_chg": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Asset Chg"
          },
          "inflow": {
            "type": "integer",
            "title": "Inflow"
          },
          "price_effect": {
            "type": "integer",
            "title": "Price Effect"
          },
          "residual": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residual"
          },
          "n_checked": {
            "type": "integer",
            "title": "N Checked"
          },
          "n_failed": {
            "type": "integer",
            "title": "N Failed"
          },
          "n_unavailable": {
            "type": "integer",
            "title": "N Unavailable"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "net_asset_chg",
          "inflow",
          "price_effect",
          "residual",
          "n_checked",
          "n_failed",
          "n_unavailable"
        ],
        "title": "Check3",
        "description": "검산 ③(순자산 변화 = 순유입 + 가격효과) — 집계에 넣은 흐름의 합."
      },
...
      "CsrfResponse": {
        "properties": {
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "csrf_token"
        ],
        "title": "CsrfResponse"
      },
...
      "Envelope_EtfFlows_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/EtfFlows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[EtfFlows]"
      },
      "Envelope_EtfHoldingChanges_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/EtfHoldingChanges"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[EtfHoldingChanges]"
      },
      "Envelope_EtfPremium_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/EtfPremium"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[EtfPremium]"
      },
      "Envelope_EtfTypes_": {
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "data": {
            "$ref": "#/components/schemas/EtfTypes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "as_of",
          "quality",
          "generated_at",
          "data"
        ],
        "title": "Envelope[EtfTypes]"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorBody",
        "description": "오류 응답(401·403·413·415·503) — 사유 코드만, 입력값·내부 문구 없음."
      },
...
      "EtfFlows": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "in",
              "out",
              "value",
              "indiv",
              "foreign",
              "inst"
            ],
            "title": "Mode"
          },
          "etf_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EtfType"
              },
              {
                "type": "null"
              }
            ]
          },
          "period": {
            "type": "integer",
            "title": "Period"
          },
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "n_total": {
            "type": "integer",
            "title": "N Total"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EtfFlowRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "new_listings": {
            "items": {
              "$ref": "#/components/schemas/NewListing"
            },
            "type": "array",
            "title": "New Listings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "etf_type",
          "period",
          "start",
          "end",
          "n_total",
          "rows",
          "new_listings"
        ],
        "title": "EtfFlows"
      },
      "EtfHoldingChanges": {
        "properties": {
          "run_date": {
            "type": "string",
            "format": "date",
            "title": "Run Date"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ChangeRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_date",
          "rows"
        ],
        "title": "EtfHoldingChanges"
      },
      "EtfPremium": {
        "properties": {
          "basis": {
            "type": "string",
            "enum": [
              "nav",
              "inav"
            ],
            "title": "Basis"
          },
          "n_checked": {
            "type": "integer",
            "title": "N Checked"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PremiumRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "basis",
          "n_checked",
          "rows"
        ],
        "title": "EtfPremium"
      },
      "EtfType": {
        "type": "string",
        "enum": [
          "kr_index",
          "kr_theme",
          "overseas",
          "leveraged_inverse",
          "bond_cash",
          "commodity",
          "other"
        ],
        "title": "EtfType",
        "description": "ETF 유형(docs/metrics.md §8 정본 — 규칙은 `kbj.engines.etf.types.etf_type`).\n\n값은 DB `prv_etf.meta.etf_type` 에 그대로 들어간다. 화면 이름은 `label`."
      },
      "EtfTypes": {
        "properties": {
          "period": {
            "type": "integer",
            "title": "Period"
          },
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/TypeRollup"
            },
            "type": "array",
            "title": "Rows"
          },
          "check3": {
            "$ref": "#/components/schemas/Check3"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period",
          "start",
          "end",
          "rows",
          "check3"
        ],
        "title": "EtfTypes"
      },
...
      "LoginBody": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginBody",
        "description": "POST /api/auth/login 본문. 값은 검증 오류 문구에 싣지 않는다(app 의 422 처리기)."
      },
...
      "MeResponse": {
        "properties": {
          "user": {
            "type": "string",
            "title": "User"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "csrf_token"
        ],
        "title": "MeResponse"
      },
      "NewListing": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "listed_on": {
            "type": "string",
            "format": "date",
            "title": "Listed On"
          },
          "net_asset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Asset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "listed_on",
          "net_asset"
        ],
        "title": "NewListing"
      },
      "NoDataBody": {
        "properties": {
          "code": {
            "type": "string",
            "const": "no_data",
            "title": "Code",
            "default": "no_data"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "NoDataBody",
        "description": "데이터가 아직 없을 때(404) — 화면은 '아직 없음 — <작업> <예정 시각>' 으로 그린다(§6.7)."
      },
...
      "Quality": {
        "type": "string",
        "enum": [
          "ok",
          "stale",
          "estimated",
          "invalid"
        ],
        "title": "Quality",
        "description": "값의 품질. 문자열 값이 그대로 DB `quality` 열과 API 응답에 들어간다."
      },
...
      "TypeRollup": {
        "properties": {
          "etf_type": {
            "$ref": "#/components/schemas/EtfType"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "net_inflow": {
            "type": "integer",
            "title": "Net Inflow"
          },
          "price_effect": {
            "type": "integer",
            "title": "Price Effect"
          },
          "n_etfs": {
            "type": "integer",
            "title": "N Etfs"
          },
          "n_new": {
            "type": "integer",
            "title": "N New"
          },
          "n_invalid": {
            "type": "integer",
            "title": "N Invalid"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "etf_type",
          "label",
          "net_inflow",
          "price_effect",
          "n_etfs",
          "n_new",
          "n_invalid"
        ],
        "title": "TypeRollup"
      }
    }
  }
}